testing: Move AbstractTransportTest to core to avoid Truth dep
Fixes #5301
