+++ core/src/test/java/io/grpc/inprocess/InProcessTransportTest.java
@@ -18,10 +18,10 @@ package io.grpc.inprocess;
 
 import com.google.common.collect.ImmutableList;
 import io.grpc.ServerStreamTracer;
+import io.grpc.internal.AbstractTransportTest;
 import io.grpc.internal.GrpcUtil;
 import io.grpc.internal.InternalServer;
 import io.grpc.internal.ManagedClientTransport;
-import io.grpc.internal.testing.AbstractTransportTest;
 import java.util.List;
 import org.junit.Ignore;
 import org.junit.Test;
