+++ testing/build.gradle
@@ -8,13 +8,6 @@ dependencies {
         // prefer 1.3 from JUnit instead of 1.1
         exclude group: 'org.hamcrest', module: 'hamcrest-core'
     }
-    // Use compileOnly to avoid dependencyConvergence problem with the Guava
-    // pulled in via Truth, for users that don't use Truth. Truth requires a
-    // more up-to-date Guava than we support elsewhere, which would trigger
-    // convergence failures in tests that only our users could resolve. Using
-    // compileOnly means only users using Truth would have the problem and
-    // they'd have to resolve it like normal anyway.
-    compileOnly libraries.truth
 
     testCompile project(':grpc-testing-proto'),
             project(':grpc-core').sourceSets.test.output
